app/google_auth/requirements.txt
google-auth-oauthlib==1.0.0
google-auth==2.17.3
google-api-python-client==2.86.0
psycopg2-binary==2.9.9
requests==2.31.0
Flask==2.3.2
Flask-Session==0.5.0
PyJWT==2.7.0
flask_cors==3.0.10 
